chapter07/5: Merge apiconv and api
- Merge apiconv and api into internal/api package
- Rename ToAPIResponse --> ToGameResponse
- Move content from game.go, guess.go into http.go
