--- 07-http_gordle/5_play/api/doc.go
@@ -1,2 +1,0 @@
-// Package api exposes paths and structures to interact with the Gordle server.
-package api
