--- 07-http_gordle/5_play/api/guess.go
@@ -1,7 +1,0 @@
-package api
-
-// GuessRequest is the structure of the message used when submitting a guess.
-type GuessRequest struct {
-	// Guess is a word attempted by the player.
-	Guess string `json:"guess"`
-}
