+++ 07-http_gordle/5_play/internal/api/convert.go
@@ -1,18 +1,17 @@
-package apiconversion
+package api
 
 import (
-	"learngo-pockets/httpgordle/api"
 	"learngo-pockets/httpgordle/internal/session"
 )
 
-// ToAPIResponse converts a domain.Game into an api.Response.
-func ToAPIResponse(g session.Game) api.GameResponse {
+// ToGameResponse converts a domain.Game into an api.Response.
+func ToGameResponse(g session.Game) GameResponse {
 	solution := g.Gordle.ShowAnswer()
 
-	apiGame := api.GameResponse{
+	apiGame := GameResponse{
 		ID:           string(g.ID),
 		AttemptsLeft: g.AttemptsLeft,
-		Guesses:      make([]api.Guess, len(g.Guesses)),
+		Guesses:      make([]Guess, len(g.Guesses)),
 		Status:       string(g.Status),
 		WordLength:   byte(len(solution)),
 	}
