+++ 07-http_gordle/5_play/internal/api/convert_test.go
@@ -1,16 +1,15 @@
-package apiconversion
+package api_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 
-	"learngo-pockets/httpgordle/api"
-	"learngo-pockets/httpgordle/internal/gordle"
+	"learngo-pockets/httpgordle/internal/api"
 	"learngo-pockets/httpgordle/internal/session"
 )
 
-func TestToAPIResponse(t *testing.T) {
+func TestToGameResponse(t *testing.T) {
 	id := "1682279480"
 	tt := map[string]struct {
 		game session.Game
@@ -18,11 +17,7 @@ func TestToAPIResponse(t *testing.T) {
 	}{
 		"nominal": {
 			game: session.Game{
-				ID: session.GameID(id),
-				Gordle: func() gordle.Game {
-					g, _ := gordle.New("HELLO")
-					return *g
-				}(),
+				ID:           session.GameID(id),
 				AttemptsLeft: 4,
 				Guesses: []session.Guess{{
 					Word:     "FAUNE",
@@ -37,16 +32,15 @@ func TestToAPIResponse(t *testing.T) {
 					Word:     "FAUNE",
 					Feedback: "⬜️🟡⬜️⬜️⬜️",
 				}},
-				WordLength: 5,
-				Solution:   "",
-				Status:     session.StatusPlaying,
+				Solution: "",
+				Status:   session.StatusPlaying,
 			},
 		},
 	}
 
 	for name, tc := range tt {
 		t.Run(name, func(t *testing.T) {
-			got := ToAPIResponse(tc.game)
+			got := api.ToGameResponse(tc.game)
 			assert.Equal(t, tc.want, got)
 		})
 	}
