+++ 07-http_gordle/5_play/internal/api/doc.go
@@ -1,0 +1,2 @@
+// Package api holds paths and structures to interact with the Gordle server.
+package api
