--- 07-http_gordle/5_play/internal/handlers/apiconversion/doc.go
@@ -1,2 +1,0 @@
-// Package apiconversion is in charge of conversion of API messages to and from domain messages.
-package apiconversion
