+++ 07-http_gordle/5_play/internal/handlers/getstatus/handler.go
@@ -8,8 +8,7 @@ import (
 
 	"github.com/go-chi/chi/v5"
 
-	"learngo-pockets/httpgordle/api"
-	"learngo-pockets/httpgordle/internal/handlers/apiconversion"
+	"learngo-pockets/httpgordle/internal/api"
 	"learngo-pockets/httpgordle/internal/repository"
 	"learngo-pockets/httpgordle/internal/session"
 )
@@ -40,7 +39,7 @@ func Handler(db gameFinder) http.HandlerFunc {
 			return
 		}
 
-		apiGame := apiconversion.ToAPIResponse(game)
+		apiGame := api.ToGameResponse(game)
 
 		w.Header().Set("Content-Type", "application/json")
 		err = json.NewEncoder(w).Encode(apiGame)
