+++ 07-http_gordle/5_play/internal/handlers/guess/handler.go
@@ -9,9 +9,8 @@ import (
 
 	"github.com/go-chi/chi/v5"
 
-	"learngo-pockets/httpgordle/api"
+	"learngo-pockets/httpgordle/internal/api"
 	"learngo-pockets/httpgordle/internal/gordle"
-	"learngo-pockets/httpgordle/internal/handlers/apiconversion"
 	"learngo-pockets/httpgordle/internal/repository"
 	"learngo-pockets/httpgordle/internal/session"
 )
@@ -54,7 +53,7 @@ func Handler(db gameGuesser) http.HandlerFunc {
 			return
 		}
 
-		apiGame := apiconversion.ToAPIResponse(game)
+		apiGame := api.ToGameResponse(game)
 
 		w.Header().Set("Content-Type", "application/json")
 		err = json.NewEncoder(w).Encode(apiGame)
